requirement.txt
python Modules:

tkinter(tk)
openCV2(cv2)
pillow(PIL)
tensorflow 2.5
pygame
mideapipe
